# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BIKO API",
        "description": "API para gerenciamento de usuários, categorias e publicações",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://biko-api.test/api",
            "description": "Servidor da API"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastrar novo usuário",
                "description": "Cria um novo usuário e retorna token de autenticação",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "city_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fulano de tal"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "fulano@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "city_id": {
                                        "description": "ID da cidade (cities.id)",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cadastro realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuário",
                "description": "Realiza login e retorna token de autenticação",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "fulano@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout do usuário autenticado",
                "description": "Invalida o token atual do usuário",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Detalhar categoria pelo slug",
                "description": "Retorna os dados de uma categoria específica usando o slug.",
                "operationId": "f9287689f645443ba435d93df47f30ab",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Slug da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "eletricista"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria encontrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Eletricista"
                    },
                    "slug": {
                        "type": "string",
                        "example": "eletricista"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Informe o token no formato: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}